set data type

+++ lab4/poligonos/Poligono.java
@@ -3,7 +3,7 @@ package lab4.poligonos;
 interface Poligono {
     int tamanhoCanvas = 100;
 
-    public float calculaArea();
+    public double calculaArea();
 
     public void imprimeTipoPoligono();
 }
